Refactored the view logic and split out controller for auth
This is the first part of a several commit change where I'm going to take the oldest part of koil (the sign-up/login flows) and apply learnings we've made since starting.
