+++ src/webapp/cypress/integration/public/index.js
@@ -1,0 +1,5 @@
+describe(` Index page`, () => {
+    it('should load successsfully', () => {
+        cy.visit("/")
+    })
+})
