+++ src/webapp/cypress/integration/public/sign-up.js
